+++ Common/src/main/java/fuzs/puzzleslib/impl/content/client/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package fuzs.puzzleslib.impl.content.client;
+
+import org.jetbrains.annotations.ApiStatus;